+++ Source/GameLoop/Game.cpp
@@ -14,20 +14,18 @@
 
 #include <chrono>
 
-
-Game::Game() {
-	window = nullptr;
-	ui = nullptr;
-}
-
 void Game::Start() {
 	MethaneVersion();
+
 	SDL_Init(SDL_INIT_VIDEO | SDL_INIT_AUDIO);
-	window = SDL_CreateWindow("Window", SDL_WINDOWPOS_CENTERED, SDL_WINDOWPOS_CENTERED, 
-		Global::windowWidth, Global::windowHeight, SDL_WINDOW_OPENGL | SDL_WINDOW_SHOWN);
+
+	Global::windowWidth = 1280;
+	Global::windowHeight = 720;
+
+	window.Init("Window", SDL_WINDOWPOS_CENTERED, SDL_WINDOWPOS_CENTERED, Global::windowWidth, Global::windowHeight, SDL_WINDOW_OPENGL | SDL_WINDOW_SHOWN);
 
 	renderer = new MT::Renderer();
-	renderer->Start(window, MT::Init(window));
+	renderer->Start(window);
 
 	Global::defaultDrawColor[0] = 255;
 	Global::defaultDrawColor[1] = 255;
@@ -46,85 +44,19 @@ void Game::Start() {
 
 	renderer->FLatRenderCopySetUp();
 
-	constexpr const char* vertexStr = R"glsl(
-        #version 330 core
-        layout(location = 0) in vec4 aRect;
-        layout(location = 1) in vec4 aSourceRect;
-        layout(location = 2) in float aAlpha;
+	FileExplorer fe;
+	fe.Open();
 
-        out vec4 oTexCoord;
-        out float oAlpha;
+	PathMaker pe;
+	pe.Open();
 
-        uniform vec2 uVievPort;
+	Label* lb = ui->CreateLabel("Test1",100,100,100,100,nullptr,ui->GetFont("arial12"),"Some text");
+	lb->SetColor(30, 30, 30);
 
-        const vec2 uvs[6] = vec2[6](
-            vec2(0.0, 0.0), // 0
-            vec2(0.0, 1.0), // 1
-            vec2(1.0, 0.0), // 2
-            vec2(0.0, 1.0), // 3
-            vec2(1.0, 1.0), // 4
-            vec2(1.0, 0.0)  // 5
-        );
+	lb = ui->CreateLabel("Test2", 100, 400, 100, 100, nullptr, ui->GetFont("arial12"), "Some text");
+	lb->SetColor(255, 30, 30);
+	lb->SetHoverFilter(true, 255, 255, 255, 120);
 
-        vec2 indexPos[6] = vec2[](
-            vec2(0.0, -1.0), // Left Down
-            vec2(0.0, 0.0),  // Left Up
-            vec2(1.0, -1.0), // Right Down
-            vec2(0.0, 0.0),  // Left Up
-            vec2(1.0, 0.0),  // Right Up
-            vec2(1.0, -1.0)  // Right Down
-        );
-
-        void main() {
-            float ndcX = (aRect.x / uVievPort.x) * 2.0 - 1.0;
-            float ndcY = 1.0 - (aRect.y / uVievPort.y) * 2.0;
-            float ndcW = (aRect.z / uVievPort.x) * 2.0;
-            float ndcH = (aRect.w / uVievPort.y) * 2.0;
-
-	        vec2 localPos = indexPos[gl_VertexID % 6];
-
-            vec2 finalPos = vec2(ndcX + localPos.x * ndcW, ndcY + localPos.y * ndcH);
-
-            gl_Position = vec4(finalPos, 0.0, 1.0);
-
-            vec2 texUvs = uvs[gl_VertexID % 6];
-
-            float u = aSourceRect.x + texUvs.x * aSourceRect.z;
-    
-            float v0 = 1.0 - aSourceRect.y - aSourceRect.w; 
-            float v = v0 + texUvs.y * aSourceRect.w;        
-
-            oTexCoord = vec4(vec2(u, v),uvs[gl_VertexID % 6]);
-            oAlpha = aAlpha;
-        }
-        )glsl";
-
-	constexpr const char* fragmentStr = R"glsl(
-        #version 330 core
-
-        out vec4 FragColor;
-
-        in vec4 oTexCoord;
-        in float oAlpha;
-
-        uniform sampler2D texture1;
-        uniform sampler2D texture2;
-
-
-        void main(){
-	        vec4 texcolor = texture(texture1,oTexCoord.xy);
-	        vec4 texcolor2 = texture(texture2,oTexCoord.zw);
-	        texcolor.a *= oAlpha;
-	        FragColor = vec4(texcolor.r * texcolor2.r, texcolor.g * texcolor2.g, 
-		        texcolor.b * texcolor2.b, texcolor.a);
-        }
-        )glsl";
-
-	cShader.Create(renderer, vertexStr, fragmentStr, {4,4,1});
-	cShader.AddUniform("uVievPort");
-
-	unsigned int vievPort = cShader.GetUnifrom("uVievPort");
-	glUniform2f(vievPort, Global::windowWidth, Global::windowHeight);
 
 }
 
@@ -141,6 +73,24 @@ void Game::FrameUpdate() {
 void Game::Input() {
 	while (SDL_PollEvent(&event)) {
 		ui->ManageInput(event);
+		if (event.type == SDL_KEYUP && event.key.keysym.scancode == SDL_SCANCODE_F) {
+			window.ToggleFullScreen();
+			Point size = window.GetSize();
+			std::println("Window Size W: {} H: {}", size.x, size.y);
+			renderer->Resize(size.x, size.y);
+		}
+		if (event.type == SDL_KEYUP && event.key.keysym.scancode == SDL_SCANCODE_R) {
+			window.Resize(300,300);
+			Point size = window.GetSize();
+			std::println("Window Size W: {} H: {}", size.x, size.y);
+			renderer->Resize(size.x, size.y);
+		}
+		if (event.type == SDL_KEYUP && event.key.keysym.scancode == SDL_SCANCODE_U) {
+			window.Resize(1280, 720);
+			Point size = window.GetSize();
+			std::println("Window Size W: {} H: {}", size.x, size.y);
+			renderer->Resize(size.x, size.y);
+		}
 		Exit();
 	}
 	Global::inputDelay++;
@@ -148,11 +98,6 @@ void Game::Input() {
 
 void Game::Render() {
 	renderer->ClearFrame(Global::defaultDrawColor[0], Global::defaultDrawColor[1], Global::defaultDrawColor[2]);
-	MT::Texture* tex1 = TexMan::GetTex("tree1");
-	MT::Texture* tex2 = TexMan::GetTex("water");
-
-	cShader.Render({ 10.0f,10.0f,100.0f,100.0f,  0.0f,0.0f, 1.0f, 1.0f, 255.0f },tex1, tex2);
-	renderer->RenderCopy({ 200,200,200,200 }, tex1);
 	ui->Render();
 	renderer->Present();
 
@@ -174,6 +119,6 @@ Game::~Game() {
 	SceneMan::Clear();
 	renderer->Clear();
 	ui->ClearAll();
-	SDL_DestroyWindow(window);
+
 	SDL_Quit();
 }